@@ -99,28 +99,28 @@
   },
   "private": false,
   "devDependencies": {
-    "@angular-devkit/build-angular": "19.2.15",
-    "@angular-devkit/core": "19.2.9",
-    "@angular-devkit/schematics": "19.2.9",
-    "@angular-eslint/eslint-plugin": "19.2.0",
-    "@angular-eslint/eslint-plugin-template": "19.2.0",
-    "@angular-eslint/template-parser": "19.2.0",
-    "@angular/cli": "~19.2.0",
-    "@angular/compiler-cli": "19.2.11",
-    "@angular/language-service": "19.2.11",
+    "@angular-devkit/build-angular": "20.0.5",
+    "@angular-devkit/core": "20.0.5",
+    "@angular-devkit/schematics": "20.0.5",
+    "@angular-eslint/eslint-plugin": "20.0.0",
+    "@angular-eslint/eslint-plugin-template": "20.0.0",
+    "@angular-eslint/template-parser": "20.0.0",
+    "@angular/cli": "~20.0.0",
+    "@angular/compiler-cli": "20.0.6",
+    "@angular/language-service": "20.0.6",
     "@ngrx/eslint-plugin": "19.0.0",
-    "@nx/angular": "21.0.3",
-    "@nx/cypress": "21.0.3",
-    "@nx/eslint": "21.0.3",
-    "@nx/eslint-plugin": "21.0.3",
-    "@nx/jest": "21.0.3",
-    "@nx/workspace": "21.0.3",
-    "@schematics/angular": "19.2.9",
+    "@nx/angular": "21.2.2",
+    "@nx/cypress": "21.2.2",
+    "@nx/eslint": "21.2.2",
+    "@nx/eslint-plugin": "21.2.2",
+    "@nx/jest": "21.2.2",
+    "@nx/workspace": "21.2.2",
+    "@schematics/angular": "20.0.5",
     "@types/jest": "^29.5.13",
     "@types/node": "^20.16.11",
-    "@typescript-eslint/eslint-plugin": "8.19.0",
-    "@typescript-eslint/parser": "8.19.0",
-    "@typescript-eslint/utils": "8.19.0",
+    "@typescript-eslint/eslint-plugin": "8.29.0",
+    "@typescript-eslint/parser": "8.29.0",
+    "@typescript-eslint/utils": "8.29.0",
     "autoprefixer": "^10.4.20",
     "cypress": "14.2.1",
     "eslint": "^8.57.1",
@@ -130,9 +130,9 @@
     "jest": "29.7.0",
     "jest-environment-jsdom": "29.7.0",
     "jest-junit": "^16.0.0",
-    "jest-preset-angular": "14.4.2",
-    "ng-packagr": "19.2.2",
-    "nx": "21.0.3",
+    "jest-preset-angular": "14.6.0",
+    "ng-packagr": "20.0.1",
+    "nx": "21.2.2",
     "openapi-mermaid": "^3.346.1",
     "openapi-ts-generator": "^8.115.3",
     "postcss": "^8.5.3",
@@ -142,19 +142,19 @@
     "prettier": "^3.5.3",
     "ts-jest": "^29.3.4",
     "ts-node": "^10.9.2",
-    "typescript": "5.7.3"
+    "typescript": "5.8.3"
   },
   "dependencies": {
-    "@angular/animations": "19.2.11",
-    "@angular/common": "19.2.11",
-    "@angular/compiler": "19.2.11",
-    "@angular/core": "19.2.11",
-    "@angular/forms": "19.2.11",
-    "@angular/localize": "20.0.0",
-    "@angular/platform-browser": "19.2.11",
-    "@angular/platform-browser-dynamic": "19.2.11",
-    "@angular/router": "19.2.11",
-    "@angular/service-worker": "19.2.11",
+    "@angular/animations": "20.0.6",
+    "@angular/common": "20.0.6",
+    "@angular/compiler": "20.0.6",
+    "@angular/core": "20.0.6",
+    "@angular/forms": "20.0.6",
+    "@angular/localize": "20.0.6",
+    "@angular/platform-browser": "20.0.6",
+    "@angular/platform-browser-dynamic": "20.0.6",
+    "@angular/router": "20.0.6",
+    "@angular/service-worker": "20.0.6",
     "@ericblade/quagga2": "^1.7.0",
     "@microsoft/applicationinsights-web": "^3.3.3",
     "@microsoft/signalr": "^8.0.7",
