@@ -11,7 +11,7 @@ export function oidcConfigurator(
       ...oidcClientConfig,
       extraQueryParams: { audience: oidcClientConfig.audience },
       redirect_uri: `${document.location.origin}/callback.html`,
-      response_type: oidcClientConfig.response_type || 'id_token token',
+      response_type: oidcClientConfig.response_type ?? 'id_token token',
       post_logout_redirect_uri: `${document.location.origin}/signout-callback.html`,
       silent_redirect_uri: `${document.location.origin}/renew-callback.html`,
       automaticSilentRenew: isNullOrUndefined(
@@ -19,7 +19,9 @@ export function oidcConfigurator(
       )
         ? true
         : oidcClientConfig.automaticSilentRenew,
-      metadataUrl: oidcClientConfig.metadataUrl || `${oidcClientConfig.authority}/.well-known/openid-configuration`,
+      metadataUrl:
+        oidcClientConfig.metadataUrl ??
+        `${oidcClientConfig.authority}/.well-known/openid-configuration`,
     },
     getUserMetadata: oidcClientConfig.getUserMetadata,
     useCallbackFlag: isNullOrUndefined(oidcClientConfig.useCallbackFlag)
